Makefile: Fix bugs and replace non-portable stuff
Add main.o to LDLIBS, to get it linked to everything.

Remove substitutions over SRCS, because if the substitution fails it
could be "dangerous" with the cleaning targets.

Remove usage of $@ in dependencies, it doesn't work in GNU make.

Sadly making the objects compile individually requires non-portable
constructs; the only alternative is to compile from the .c files
directly, which is what the default rules do.

Signed-off-by: Ismael Luceno <ismael.luceno@tttech-auto.com>